+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Added
 - Package and deploy releases to npm.js as [lua-stuart](https://www.npmjs.com/package/lua-stuart).
 
+### Fixed
+- [Fengari Web](https://github.com/fengari-lua/fengari-web) interop fails because Stuart assumes an 'io' module is always present for logging
+
 ## [0.1.7] - 2018-09-15
 ### Added
 - Dockerfiles provided for testing with specific versions of Lua 5.1, 5.2, 5.3